+++ tests/bootstrap.php
@@ -12,7 +12,6 @@ require_once(__DIR__.'/../vendor/autoload.php');
 
 // Configure VCR
 ini_set('opcache.enable', '0');
-VCR::configure()->setCassettePath(__DIR__.'/fixtures/VCR/');
 VCR::configure()->setMode(VCR::MODE_ONCE);
 VCR::configure()->enableLibraryHooks(['stream_wrapper', 'curl'])
     ->addRequestMatcher('headers_custom_matcher', static function (Request $first, Request $second) : bool {